rpi-default-providers: Document how to switch providers for vc4 gfx stack
The Raspberry Pi boards can use one of two graphics stacks: The "userland"
user-space driver or the VC4 drm/kms kernel driver. This patch documents
how the user user can choose if the VC4 driver has to be used instead and
set the right preferred providers for that case.

Signed-off-by: Derek Foreman <derekf@osg.samsung.com>
[javier: Extend commit msg and made conditional instead removing userland]
Signed-off-by: Javier Martinez Canillas <javier@osg.samsung.com>

Series-changes: 2
- Don't add a feature and allow users to change the default provider manually
  aided by a comment. Suggested by Andreas Müller and Andrei Gherzan.

Series-version: 2

Series-to: Andrei Gherzan <andrei@gherzan.ro>
Series-cc: yocto@yoctoproject.org
Series-cc: Andreas Müller <schnitzeltony@googlemail.com>
Series-cc: Khem Raj <raj.khem@gmail.com>
Series-cc: Petter Mabäcker <petter@technux.se>
Series-cc: Derek Foreman <derekf@osg.samsung.com>
Series-cc: Mauro Carvalho Chehab <mchehab@osg.samsung.com>

Cover-letter:
Add support for 4.1 kernel with vc4 DRM/KMS driver
Hello Andrei,

This series adds support for Eric Anholt's v4.1 kernel, that has support
for the vc4 DRM/KMS driver. Which is the new open source graphics driver
stack for the Raspberry Pi to be used instead of the userland driver.

This is a second version of the series which addresses feedback provided
by Andreas Müller, Khem Raj, Petter Mabäcker and you. The first version
of the series can be found in [0]. Some of the v1 patches have already
been picked and are in the meta-raspberrypi master branch.

The v4.1 kernel is under heavy development so is a work-in-progress and
should not be used in production. That's why is not set as the default
virtual/linux provider but the user can change the default providers in
case wants to user the VC4 driver.

But even when it's still a development kernel, having the recipe in the
meta-raspberrypi will allow people to test it. We have been using it in
the Tizen RPi2 port [1]. The patches in the series are for:

Patch 1/4 allows to set the mask_gpu_interrupt0 option in config.txt

Patch 2/4 adds a recipe for the 4.1 and some patches to make it stable

Patch 3/4 Documents how to switch the default providers for vc4 gfx stack

Patch 4/4 Adds a section to the README that explains that two gfx stacks
are available and that the user can change the default providers.

As discussed in v1, the following has to be added in local.conf to build:

KERNEL_DEVICETREE = " \
    bcm2708-rpi-b.dtb \
    bcm2708-rpi-b-plus.dtb \
    bcm2709-rpi-2-b.dtb \
    \
    overlays/hifiberry-amp-overlay.dtb \
    overlays/hifiberry-dac-overlay.dtb \
    overlays/hifiberry-dacplus-overlay.dtb \
    overlays/hifiberry-digi-overlay.dtb \
    overlays/iqaudio-dac-overlay.dtb \
    overlays/iqaudio-dacplus-overlay.dtb \
    overlays/lirc-rpi-overlay.dtb \
    overlays/pps-gpio-overlay.dtb \
    overlays/w1-gpio-overlay.dtb \
    overlays/w1-gpio-pullup-overlay.dtb \
    "

[0]: https://www.mail-archive.com/yocto@yoctoproject.org/msg25164.html
[1]: http://blogs.s-osg.org/tizen-rpi2-now-supporting-3d-acceleration/

Best regards,
Javier
END