rpi-config: Allow to mask GPU irqs
The rpi firmware has support for a mask_gpu_interrupt option that can be
defined in the boot/config.txt file. The option makes the VPU not handle
certain interrupts such as V3D. This option is needed by the VC4 DRM/KMS
driver but it is not possible to set it from the meta-raspberrypi layer.

Signed-off-by: Derek Foreman <derekf@osg.samsung.com>
[javier: Made it a config option]
Signed-off-by: Javier Martinez Canillas <javier@osg.samsung.com>

Series-changes: 2
 - Fix commit message about mask GPU irq option. Suggested by Andrei Gherzan.