linux-raspberrypi: Add a 4.1 linux kernel with vc4 support
This adds Eric Anholt's WIP kernel with dri/kms/3d support for the GPU on
the rpi2. Adding a recipe that tracks this kernel, will make it easier to
test this setup.

This recipe should only used for testing purposes for now, so it will not
be set as the default.

The kernel will be choosen though if the "vc4-gfx" feature has been added
to the the DISTRO_FEATURES variable.

Signed-off-by: Derek Foreman <derekf@osg.samsung.com>
[javier: Extended commit message and set default preference to -1]
Signed-off-by: Javier Martinez Canillas <javier@osg.samsung.com>